fix: refactor Card, CardList, GameCard, GameCardList
