+++ src/App.tsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { createBrowserRouter, Outlet, RouterProvider } from 'react-router-dom';
-import Home from './routes/Home';
-import NotFound from './routes/NotFound';
-import AboutUs from './routes/AboutUs';
-import NavBar from './routes/NavBar';
-import Form from './routes/Form';
+import Home from './routes/Home/Home';
+import NotFound from './routes/NotFound/NotFound';
+import AboutUs from './routes/AboutUs/AboutUs';
+import NavBar from './routes/NavBar/NavBar';
+import Form from './routes/Form/Form';
 
 const router = createBrowserRouter([
   {
