+++ src/components/FormComponents/TextInput.tsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
-import { InputProps } from '../types/InputProps';
+import { InputProps } from '../../types/InputProps';
+import { GameCardData } from '../GameCard/GameCard';
 
-class DateInput extends Component<InputProps> {
+class TextInput extends Component<InputProps> {
   render() {
     const { label, inputRef, name, errors, ...attributes } = this.props;
     return (
@@ -9,16 +10,18 @@ class DateInput extends Component<InputProps> {
         <label className="control__label" htmlFor={name}>{`${label}:`}</label>
         <input
           id={name}
-          type="date"
           ref={inputRef}
-          className="control__input"
+          className="control__input text__input"
           aria-label={name}
+          type="text"
           {...attributes}
         />
-        {errors.firstRelease && <span className="error-msg">{errors.firstRelease}</span>}
+        {errors[name as keyof GameCardData] && (
+          <span className="error-msg">{errors[name as keyof GameCardData]}</span>
+        )}
       </div>
     );
   }
 }
 
-export default DateInput;
+export default TextInput;
